Revert change

+++ app/controllers/application_controller.rb
@@ -14,11 +14,12 @@ class ApplicationController < ActionController::Base
   end
 
   def current_user
-    @current_user ||= if session[:user_id]
-                        User.find_by(id: session[:user_id]).tap do |user|
-                          user ? gon.push(user_id: user.id, team_id: user.team_id) : session.delete(:user_id)
-                        end
-                      end
+    @current_user ||= begin
+      if session[:user_id]
+        User.find_by(id: session[:user_id]).tap do |user|
+          user ? gon.push(user_id: user.id, team_id: user.team_id) : session.delete(:user_id)
+      end
+    end
   end
   helper_method :current_user
